ST-DBSCAN
This program takes in a file containing latitude, longitude of different
places along with area code and NDVI values and gives as output the
clusters based on pre-defined value of epsilon. Heuristic approach to
determining epsilon can be implemented using slope of graph but I have
not used it here.

For more information on ST-DBSCAN algorithm, refer this:
http://www.sciencedirect.com/science/article/pii/S0169023X06000218